requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
brotli-python=1.0.9=py311h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2023.11.17=py311h06a4308_0
cffi=1.16.0=py311h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py311h06a4308_0
cryptography=41.0.7=py311hdda0065_0
greenlet=3.0.1=py311h6a678d5_0
idna=3.4=py311h06a4308_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=3.0.12=h7f8727e_0
pip=23.3.1=py311h06a4308_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.2.0=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
python-dotenv=0.21.0=py311h06a4308_0
readline=8.2=h5eee18b_0
requests=2.31.0=py311h06a4308_0
setuptools=68.2.2=py311h06a4308_0
sqlalchemy=2.0.21=py311h5eee18b_0
sqlite=3.41.2=h5eee18b_0
tk=8.6.12=h1ccaba5_0
typing-extensions=4.7.1=py311h06a4308_0
typing_extensions=4.7.1=py311h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py311h06a4308_0
wheel=0.41.2=py311h06a4308_0
xz=5.4.5=h5eee18b_0
zlib=1.2.13=h5eee18b_0
